blog_proj

+++ blog_proj/src/main/java/com/kosta/service/BlogServiceImpl.java
@@ -6,29 +6,28 @@ import java.util.Optional;
 import org.springframework.stereotype.Service;
 
 import com.kosta.entity.Article;
+import com.kosta.entity.User;
 import com.kosta.repository.BlogRepository;
 
 import lombok.RequiredArgsConstructor;
 
 @Service
-@RequiredArgsConstructor // 생성자 생성, @autowired 대체
+@RequiredArgsConstructor
 public class BlogServiceImpl implements BlogService {
 
-	private final BlogRepository blogRepository; // autowired 대체
+	private final BlogRepository blogRepository;
 
-	// 게시물 추가
 	@Override
-	public Article save(Article article) {		
+	public Article save(Article article, User user) {
+		article.setCreator(user);
 		return blogRepository.save(article);
 	}
 
-	// 게시물 전체 목록
 	@Override
 	public List<Article> findAll() {
 		return blogRepository.findAll();
 	}
 
-	// ID로 특정 게시물 찾기
 	@Override
 	public Article findById(Long id) throws Exception {
 		Optional<Article> optArticle = blogRepository.findById(id);
@@ -36,48 +35,40 @@ public class BlogServiceImpl implements BlogService {
 		return article;
 	}
 
-	// ID로 특정 게시물 삭제
 	@Override
-	public void deleteById(Long id) throws Exception {
+	public void deleteById(Long id, User user) throws Exception {
 		Article article = findById(id);
-		blogRepository.deleteById(article.getId()); // 없는 ID가 있는 경우 방지
+		User creator = article.getCreator();
+		// user(로그인)와 creator(작성자)의 동일 여부 분기
+		if (user.getId().equals(creator.getId())) {
+			blogRepository.deleteById(article.getId());			
+		} else {
+			throw new Exception("본인의 글만 삭제할 수 있습니다.");
+		}
 	}
 
-	// 수정
 	@Override
-	public Article update(Article article) throws Exception {
-		Article originArticle = findById(article.getId());
+	public Article update(Article article, User user) throws Exception {
+		Article originArticle = findById(article.getId());		
+		User creator = originArticle.getCreator();
+		
+		if (!user.getId().equals(creator.getId())) {
+			throw new Exception("본인의 글만 수정할 수 있습니다.");
+		}
+		
 		originArticle.setTitle(article.getTitle());
 		originArticle.setContent(article.getContent());
-		
-		Article updateArticle = blogRepository.save(originArticle);
-		return updateArticle;
+		Article updatedArticle = blogRepository.save(originArticle);
+		return updatedArticle;
 	}
 
-	// 검색 & 정렬
 	@Override
 	public List<Article> searchAndOrder(String keyword, String order) {
 		if (order.equals("desc")) {
-			// 내림 차순			
+			// 내림차순 결과 반환
 			return blogRepository.findByTitleContainsOrContentContainsOrderByTitleDesc(keyword, keyword);
-		}
-		// 오름 차순
-		return blogRepository.findByTitleContainsOrContentContainsOrderByTitleAsc(keyword, keyword);	
+		} 
+		// 오른차순 결과 반환
+		return blogRepository.findByTitleContainsOrContentContainsOrderByTitleAsc(keyword, keyword);
 	}
-
-//	// 검색
-//	@Override
-//	public List<Article> searchInTitleAndContent(String keyword) {
-//		return blogRepository.findByTitleContainsOrContentContains(keyword, keyword);
-//	}
-//	// 정렬
-//	@Override
-//	public List<Article> orderingArticleList(String order) {
-//		if (order.equals("desc")) return blogRepository.findAllByOrderByTitleDesc();		
-//		return blogRepository.findAllByOrderByTitleAsc(); // 기본 오름차순 정렬
-//	}
-	
-	
-	
-	
 }
